Fix half/float mismatches in URP shader inputs for Hybrid V2.

+++ com.unity.render-pipelines.universal/Shaders/BakedLitInput.hlsl
@@ -14,10 +14,10 @@ CBUFFER_END
 #ifdef UNITY_DOTS_INSTANCING_ENABLED
 UNITY_DOTS_INSTANCING_START(MaterialPropertyMetadata)
     UNITY_DOTS_INSTANCED_PROP(float4, _BaseMap_ST)
-    UNITY_DOTS_INSTANCED_PROP(half4 , _BaseColor)
-    UNITY_DOTS_INSTANCED_PROP(half  , _Cutoff)
-    UNITY_DOTS_INSTANCED_PROP(half  , _Glossiness)
-    UNITY_DOTS_INSTANCED_PROP(half  , _Metallic)
+    UNITY_DOTS_INSTANCED_PROP(float4, _BaseColor)
+    UNITY_DOTS_INSTANCED_PROP(float , _Cutoff)
+    UNITY_DOTS_INSTANCED_PROP(float , _Glossiness)
+    UNITY_DOTS_INSTANCED_PROP(float , _Metallic)
 UNITY_DOTS_INSTANCING_END(MaterialPropertyMetadata)
 
 #define _BaseMap_ST         UNITY_ACCESS_DOTS_INSTANCED_PROP_FROM_MACRO(float4, Metadata__BaseMap_ST)
