+++ com.unity.render-pipelines.universal/Shaders/SimpleLitInput.hlsl
@@ -15,10 +15,10 @@ CBUFFER_END
 #ifdef UNITY_DOTS_INSTANCING_ENABLED
 UNITY_DOTS_INSTANCING_START(MaterialPropertyMetadata)
     UNITY_DOTS_INSTANCED_PROP(float4, _BaseMap_ST)
-    UNITY_DOTS_INSTANCED_PROP(half4 , _BaseColor)
-    UNITY_DOTS_INSTANCED_PROP(half4 , _SpecColor)
-    UNITY_DOTS_INSTANCED_PROP(half4 , _EmissionColor)
-    UNITY_DOTS_INSTANCED_PROP(half  , _Cutoff)
+    UNITY_DOTS_INSTANCED_PROP(float4, _BaseColor)
+    UNITY_DOTS_INSTANCED_PROP(float4, _SpecColor)
+    UNITY_DOTS_INSTANCED_PROP(float4, _EmissionColor)
+    UNITY_DOTS_INSTANCED_PROP(float , _Cutoff)
 UNITY_DOTS_INSTANCING_END(MaterialPropertyMetadata)
 
 #define _BaseMap_ST         UNITY_ACCESS_DOTS_INSTANCED_PROP_FROM_MACRO(float4, Metadata__BaseMap_ST)
